feature(): refactor route run method using response helper
