+++ src/Utils/helpers.php
@@ -7,7 +7,7 @@ use Utils\Response;
 
 
 if (! function_exists('view')) {
-    function view(string $viewFile, array $opts = [])
+    function view(string $viewFile, array $opts = []): void
     {
         $template = new View();
         $template->render($viewFile, $opts);
@@ -15,7 +15,7 @@ if (! function_exists('view')) {
 }
 
 if (! function_exists('env_var')) {
-    function env_var(string $varName, $fallBackValue = '')
+    function env_var(string $varName, string $fallBackValue = '')
     {
         $envVarValue = $fallBackValue;
 
@@ -29,15 +29,15 @@ if (! function_exists('env_var')) {
 }
 
 if (! function_exists('request')) {
-    function request()
+    function request(): Request
     {
         return new Request();
     }
 }
 
 if (! function_exists('response')) {
-    function response()
+    function response(mixed $closure = null): Response
     {
-        return new Response();
+        return new Response($closure);
     }
 }
